clausewizard_backend/requirements.txt
Flask
openai
psycopg2-binary
python-dotenv
psycopg2
sentence_transformers